create CardForm component
